execution/stagedsync: move the txlookup prune suite to its own test file
no_prune_test.go is about prune entrypoints no-opping under --exec.no-prune; the
txlookup suite was 279 of its 411 lines. stage_execute_prune_test.go is the
precedent for one stage's prune entrypoint in its own file.

Pure move - same names, same bodies. It also takes freezeblocks, dbservices,
rawdbv3, config3 and crypto/sha256 out of a file that needs none of them.
